requirements.txt
kafka-python-ng
flask
